requirements.txt
black
gym
pyaml
numpy
huskarl
matplotlib
pygame